First matching and extraction algorithm implementations

+++ src/cti/ticket_image.cpp
@@ -4,7 +4,6 @@
 
 cti::TicketImage::TicketImage(const string& imagePath) : _width(0), _height(0), _bytes_per_pixel(3), _bytes_per_line(0) {
     const cv::Mat image = cv::imread(imagePath);
-    // size_t dataSize = image.total() * image.step.p[image.dims - 1];
     size_t dataSize = image.total() * image.elemSize();
 
     this->_width = image.size().width;
@@ -16,7 +15,7 @@ cti::TicketImage::TicketImage(const string& imagePath) : _width(0), _height(0), 
     memcpy(this->_image, image.data, dataSize);
 }
 
-cti::TicketImage::TicketImage(const int width, const int height, const int bytes_per_pixel, const int bytes_per_line,
+cti::TicketImage::TicketImage(const int width, const int height, const int bytes_per_pixel, const size_t bytes_per_line,
                               unsigned char *image)
         : _width(width), _height(height), _bytes_per_pixel(bytes_per_pixel), _bytes_per_line(bytes_per_line),
           _image(image) {
@@ -35,7 +34,7 @@ const int cti::TicketImage::bytesPerPixel() const {
     return _bytes_per_pixel;
 }
 
-const int cti::TicketImage::bytesPerLine() const {
+const size_t cti::TicketImage::bytesPerLine() const {
     return _bytes_per_line;
 }
